refactor: layout 컴포넌트 폴더 구조 변경 사항 적용
