+++ src/components/ui/toast/use-toaster.ts
@@ -1,9 +1,8 @@
 'use client';
 
 // Inspired by react-hot-toast library
+import type { ToastActionElement, ToastProps } from '@components/ui/toast';
 import * as React from 'react';
-
-import type { ToastActionElement, ToastProps } from '@/components/ui/toast';
 
 const TOAST_LIMIT = 1;
 const TOAST_REMOVE_DELAY = 1000000;
