+++ src/pages/council/recruitment.tsx
@@ -1,9 +1,9 @@
+import SinglePageLayout from '@components/layouts/SinglePageLayout';
 import Box from '@components/ui/box';
 import { FileBox } from '@components/ui/box/PostBox';
 import { HEADING_TEXT, HEADING_STYLE } from '@constants/heading';
 import { useEffectOnce } from '@hooks/useEffectOnce';
 import { useLayout } from '@hooks/useLayout';
-import SinglePageLayout from '@layouts/SinglePageLayout';
 import React from 'react';
 
 export default function Recruitment() {
